Updated strava API

+++ strava_V3/src/main/java/io/swagger/client/api/ActivitiesApi.java
@@ -1,5 +1,6 @@
 /*
  * Strava API v3
+ * The [Swagger Playground](https://developers.strava.com/playground) is the easiest way to familiarize yourself with the Strava API by submitting HTTP requests and observing the responses before you write any client code. It will show what a response will look like with different endpoints depending on the authorization scope you receive from your athletes. To use the Playground, go to https://www.strava.com/settings/api and change your “Authorization Callback Domain” to developers.strava.com. Please note, we only support Swagger 2.0. There is a known issue where you can only select one scope at a time. For more information, please check the section “client code” at https://developers.strava.com/docs.
  *
  * OpenAPI spec version: 3.0.0
  * 
@@ -909,7 +910,7 @@ public class ActivitiesApi {
      * @return Call to execute
      * @throws ApiException If fail to serialize the request body object
      */
-    public com.squareup.okhttp.Call getZonesByActivityIdCall(Integer id, final ProgressResponseBody.ProgressListener progressListener, final ProgressRequestBody.ProgressRequestListener progressRequestListener) throws ApiException {
+    public com.squareup.okhttp.Call getZonesByActivityIdCall(Long id, final ProgressResponseBody.ProgressListener progressListener, final ProgressRequestBody.ProgressRequestListener progressRequestListener) throws ApiException {
         Object localVarPostBody = null;
 
         // create path and map variables
@@ -952,7 +953,7 @@ public class ActivitiesApi {
     }
 
     @SuppressWarnings("rawtypes")
-    private com.squareup.okhttp.Call getZonesByActivityIdValidateBeforeCall(Integer id, final ProgressResponseBody.ProgressListener progressListener, final ProgressRequestBody.ProgressRequestListener progressRequestListener) throws ApiException {
+    private com.squareup.okhttp.Call getZonesByActivityIdValidateBeforeCall(Long id, final ProgressResponseBody.ProgressListener progressListener, final ProgressRequestBody.ProgressRequestListener progressRequestListener) throws ApiException {
         
         // verify the required parameter 'id' is set
         if (id == null) {
@@ -972,7 +973,7 @@ public class ActivitiesApi {
      * @return List&lt;ActivityZone&gt;
      * @throws ApiException If fail to call the API, e.g. server error or cannot deserialize the response body
      */
-    public List<ActivityZone> getZonesByActivityId(Integer id) throws ApiException {
+    public List<ActivityZone> getZonesByActivityId(Long id) throws ApiException {
         ApiResponse<List<ActivityZone>> resp = getZonesByActivityIdWithHttpInfo(id);
         return resp.getData();
     }
@@ -984,7 +985,7 @@ public class ActivitiesApi {
      * @return ApiResponse&lt;List&lt;ActivityZone&gt;&gt;
      * @throws ApiException If fail to call the API, e.g. server error or cannot deserialize the response body
      */
-    public ApiResponse<List<ActivityZone>> getZonesByActivityIdWithHttpInfo(Integer id) throws ApiException {
+    public ApiResponse<List<ActivityZone>> getZonesByActivityIdWithHttpInfo(Long id) throws ApiException {
         com.squareup.okhttp.Call call = getZonesByActivityIdValidateBeforeCall(id, null, null);
         Type localVarReturnType = new TypeToken<List<ActivityZone>>(){}.getType();
         return apiClient.execute(call, localVarReturnType);
@@ -998,7 +999,7 @@ public class ActivitiesApi {
      * @return The request call
      * @throws ApiException If fail to process the API call, e.g. serializing the request body object
      */
-    public com.squareup.okhttp.Call getZonesByActivityIdAsync(Integer id, final ApiCallback<List<ActivityZone>> callback) throws ApiException {
+    public com.squareup.okhttp.Call getZonesByActivityIdAsync(Long id, final ApiCallback<List<ActivityZone>> callback) throws ApiException {
 
         ProgressResponseBody.ProgressListener progressListener = null;
         ProgressRequestBody.ProgressRequestListener progressRequestListener = null;
